+++ test/BinderSpec.js
@@ -248,6 +248,16 @@ describe('Binder', function() {
     $rootScope.hidden = 'false';
     $rootScope.$apply();
 
+    assertHidden(element);
+
+    $rootScope.hidden = 0;
+    $rootScope.$apply();
+
+    assertVisible(element);
+
+    $rootScope.hidden = false;
+    $rootScope.$apply();
+
     assertVisible(element);
 
     $rootScope.hidden = '';
@@ -265,6 +275,16 @@ describe('Binder', function() {
     assertVisible(element);
 
     $rootScope.show = 'false';
+    $rootScope.$apply();
+
+    assertVisible(element);
+
+    $rootScope.show = false;
+    $rootScope.$apply();
+
+    assertHidden(element);
+
+    $rootScope.show = false;
     $rootScope.$apply();
 
     assertHidden(element);